fix multi-plugin-usage
Old version had the problem, that if multiple plugins were using the library only the migrations of the first installed (and therefor loaded) plugin were executed, since Composer only loads a library once.

To not use expensive `debug_backtrace` the library now recommends to have a Constant defined in Plugin main file and forces the plugin extending the `DBUtilisator` class to define a method `get_plugin_dir` which in the easiest way only returns this constant.